+++ main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 
-from controller.controller import router
+from src.controller.controller import router
 
 app = FastAPI()
 app.include_router(router=router)
